backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
openai>=1.30
python-dotenv>=1.0
djangorestframework-simplejwt>=5.3
django-filter>=23.5
gunicorn>=21.2